:truck: Move middleware to source
